Refactored to have correct type for flex/vaca storage.
Still failing to be correctly applied to the employee - help please?

+++ JavaSource/com/timelink/controllers/HumanResourceController.java
@@ -31,7 +31,7 @@ public class HumanResourceController implements Serializable {
   private String firstName;
   private String lastName;
   private String email;
-  private Integer vacationAccrual;
+  private Float vacationAccrual;
   private RoleEnum jobTitle;
   private String password;
   private String confirmPassword;
@@ -137,7 +137,7 @@ public class HumanResourceController implements Serializable {
    * Return the vacationAccrual.
    * @return the vacationAccrual
    */
-  public Integer getVacationAccrual() {
+  public Float getVacationAccrual() {
     return vacationAccrual;
   }
 
@@ -145,7 +145,7 @@ public class HumanResourceController implements Serializable {
    * Set the vacationAccrual to vacationAccrual.
    * @param vacationAccrual the vacationAccrual to set
    */
-  public void setVacationAccrual(Integer vacationAccrual) {
+  public void setVacationAccrual(Float vacationAccrual) {
     this.vacationAccrual = vacationAccrual;
   }
