+++ JavaSource/com/timelink/ejbs/Employee.java
@@ -79,13 +79,13 @@ public class Employee implements Serializable {
   private Date effectTo;
   
   @Column(name = "emp_flex_time")
-  private Integer flexTime;
+  private Float flexTime;
   
   @Column(name = "emp_vacation_time")
-  private Integer vacationTime;
+  private Float vacationTime;
   
   @Column(name = "emp_vacation_rate")
-  private Integer vacationRate;
+  private Float vacationRate;
   
   //private String status;
   
@@ -323,7 +323,7 @@ public class Employee implements Serializable {
    * Returns the flexTime.
    * @return the flexTime
    */
-  public int getFlexTime() {
+  public float getFlexTime() {
     return flexTime;
   }
 
@@ -331,7 +331,7 @@ public class Employee implements Serializable {
    * Sets the flexTime to flexTime.
    * @param flexTime the flexTime to set
    */
-  public void setFlexTime(int flexTime) {
+  public void setFlexTime(float flexTime) {
     this.flexTime = flexTime;
   }
 
@@ -339,7 +339,7 @@ public class Employee implements Serializable {
    * Returns the vacationTime.
    * @return the vacationTime
    */
-  public int getVacationTime() {
+  public float getVacationTime() {
     return vacationTime;
   }
 
@@ -347,7 +347,7 @@ public class Employee implements Serializable {
    * Sets the vacationTime to vacationTime.
    * @param vacationTime the vacationTime to set
    */
-  public void setVacationTime(int vacationTime) {
+  public void setVacationTime(float vacationTime) {
     this.vacationTime = vacationTime;
   }
 
@@ -355,7 +355,7 @@ public class Employee implements Serializable {
    * Returns the vacationRate.
    * @return the vacationRate
    */
-  public int getVacationRate() {
+  public float getVacationRate() {
     return vacationRate;
   }
 
@@ -363,7 +363,7 @@ public class Employee implements Serializable {
    * Sets the vacationRate to vacationRate.
    * @param vacationRate the vacationRate to set
    */
-  public void setVacationRate(int vacationRate) {
+  public void setVacationRate(float vacationRate) {
     this.vacationRate = vacationRate;
   }
